LPS-190113 Move tests inside layout_reports directory
